refactor: more restructuring
