+++ packages/vite-glob-routes/src/react-router/features/core/README.md
@@ -1,0 +1,1 @@
+core logic and utilities
